Refactor `region-manager.ts` into smaller modules
Extracted `RegionManager`'s complex LOD and spatial query logic into `region-manager-lod.ts`, moving core class definitions and key utilities to `region-manager-core.ts`. Replaced `region-manager.ts` with a barrel file to preserve backward compatibility. This improves maintainability while preserving zero-allocation paths.
